% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_change.R
\name{add_change}
\alias{add_change}
\title{Legg til geokoder endringer}
\usage{
add_change(filegeo, filechg, year, type = "land", path = NULL, raw = TRUE)
}
\arguments{
\item{filegeo}{CSV fil for alle geokoder}

\item{filechg}{Excel fil for kode endringer}

\item{year}{Året geokoder gjelder}

\item{type}{Hvilket nivå f.eks fylke, kommune osv.}

\item{path}{Sti til mappen for filene ligger}

\item{raw}{Hvis FALSE så er \code{filegeo} og \code{filechg} en ferdig
laget object i .GlobalEnv dvs. ikke er CSV og Excel fil}
}
\description{
Filen for gjeldende geokoder i \code{csv} format og filen om koder som har
endret i \code{xlsx} skal legges sammen. Disse filene hentes fra SSB
hjemmeside.
}
\details{
Følgende dataset produseres:
\itemize{
\item DT     : Hele datasettet med kode endringer
\item xl     : Endringer fra Excel filen
\item change : Sti til fil endringe hvis \code{raw = FALSE}
\item code   : Sti til kodefilen hvis \code{raw = FALSE}
}
}
