% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{cast_all}
\alias{cast_all}
\title{Add lower granularity to all files}
\usage{
cast_all(file, type, year, keep.col = c("code", "name"))
}
\arguments{
\item{file}{Files names with complete path to the files}

\item{type}{Types equivalent to the file names. Must be the same order with file names}

\item{year}{Året geokoder gjelder}

\item{keep.col}{Exisiting column names to be kept}
}
\description{
Add all granularity levels. Here the structure is important.
The the order for types should be similar to the order in files.
}
\examples{
\dontrun{
file = c("C:/Geo/fylke/fylke2020.csv",
          "C:/Geo/kommune/kommune2020.csv",
          "C:/folder2/bydel2020.csv" )

types = c("fylke","kommune","bydel")

DT <- cast_all(files=files, type=types, year=2020)
}

}
