% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{cast_code}
\alias{cast_code}
\title{Add lower granularity when applicable}
\usage{
cast_code(file, type, year, folder.path = NULL, keep.col = c("code", "name"))
}
\arguments{
\item{file}{File names}

\item{type}{Type of regional granularity ie. fylke, kommune etc.}

\item{year}{Year of the selected code list}

\item{folder.path}{Folder path to the files}

\item{keep.col}{Exisiting column names to be kept}
}
\description{
This function will add lower granularity to the current granularity when applicable.
It means that when casting a kommune file, then columns for fylke will be added.
For grunnkrets file then both kommune and fylke code will be added. This is based on
the available table that can be downloaded from SSB. The file should be in a CSV format.
}
\examples{

\dontrun{
file = "c:/Users/geo/ssb_grunnkrets_jan2020.csv"
type = "grunnkrets"
year = 2020
cols = c("code", "name")

DT <- cast_code(file = file, type = type, year = year, keep.col=cols)
}

}
