% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\docType{data}
\name{dataApi}
\alias{dataApi}
\title{Where downloaded data will be kept}
\format{
An object of class \code{environment} of length 0.
}
\usage{
dataApi
}
\description{
Downloaded data will be stored in dataApi to avoid downloding
multiple time for the same selected data
}
\keyword{datasets}
