% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_change.R
\name{find_change}
\alias{find_change}
\title{Find changes from previus year}
\usage{
find_change(newCode, preCode, raw = TRUE)
}
\arguments{
\item{newCode}{Object produced from \code{geo_set} for newer codes}

\item{preCode}{Object produced from \code{geo_set} for previous codes}

\item{raw}{If FALSE then specify a data.table object}
}
\description{
Which codes that have changed since previous codes
}
