% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_change.R
\name{geo_change}
\alias{geo_change}
\title{Get change from complete geo list}
\usage{
geo_change(
  files = NULL,
  years = NULL,
  type = NULL,
  key.col = NULL,
  folder.path = NULL,
  file.type = c("none", "xls", "csv"),
  des.path = NULL
)
}
\arguments{
\item{files}{A list of files to find geo codes changes}

\item{years}{A list of years for these files}

\item{type}{Type of regional granularity ie. fylke, kommune etc.}

\item{key.col}{Column name for merging eg. \code{name}}

\item{folder.path}{Folder path to the files}

\item{file.type}{Which format to save the output}

\item{des.path}{Destination folder where the file to be saved}
}
\description{
Create table for code changes when table tracking the changes isn't already available from SSB.
This function will be using a reference colum other than geo code ie. names. Therefore this
function should be used with caution. The output can be as \code{xlsx} or \code{csv} file.
}
\examples{

\dontrun{
files = c("ssb_bydel_jan2004.csv", "ssb_bydel_jan2018.csv", "ssb_bydel_jan2020.csv")
years = c(2004, 2018, 2020)
folder = "C:/geo/bydel"
des = "C:/geo/bydel/output"

geo_change(files=files, years=years, type="bydel",folder.path=folder, file.type="xls", des.path=des)
}

}
