% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-list.R
\name{geo_list}
\alias{geo_list}
\title{Get the list of geographical levels}
\usage{
geo_list(
  type = c("fylke", "kommune", "bydel", "grunnkrets"),
  year,
  from = NULL
)
}
\arguments{
\item{type}{Type of regional granularity ie. fylke, kommune etc.}

\item{year}{Year of the selected code list}

\item{from}{Year from}
}
\description{
This function will download the codes of selected geographical levels via API.
}
