% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{geo_merge}
\alias{geo_merge}
\title{Merge all geo codes}
\usage{
geo_merge(files, output = c("all", "change", "complete"))
}
\arguments{
\item{files}{A list of objects to be merged. The order must be from lowest to highest year.}

\item{output}{Data output which can be one of these:
\itemize{
\item \code{all}      : Dataset for all data. This is the default output
\item \code{change}   : Dataset that include only codes that have changed
\item \code{complete} : A list of all different outputs
}}
}
\description{
Create a table with all geo change and year. The files for the argument must be
the output object from \code{geo_set()} function.
}
\examples{

\dontrun{
DT <- geo_merge(list(kommune2018, kommune2019, kommune2020))
}

}
