% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_geo.R
\name{geo_save}
\alias{geo_save}
\title{Save geo codes}
\usage{
geo_save(
  tblname = NULL,
  obj = NULL,
  des.path = FALSE,
  file.type = c("Access", "SQLite", "Excel", "Text"),
  db.name = NULL
)
}
\arguments{
\item{tblname}{Name of the table to be saved as}

\item{obj}{Object name to be saved}

\item{des.path}{Destination folder where the file to be saved}

\item{file.type}{Choose file type as Access, SQLite, Excel or Text}

\item{db.name}{When choosing a DBMS then specify the database name}
}
\description{
Geo codes can be saved either in a database management system (DBMS) or
as an Excel or text file.
}
