% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-code.R
\name{get_code}
\alias{get_code}
\title{Get the codes of geographical levels}
\usage{
get_code(
  type = c("fylke", "kommune", "bydel", "grunnkrets"),
  from = NULL,
  to = NULL,
  date = FALSE
)
}
\arguments{
\item{type}{Higer granularity from specified correspond arg.}

\item{from}{Specify the starting year for range period. Current year is the default.}

\item{to}{Specify the year to end the range period. Current year is used when not specified.}

\item{date}{If TRUE then give complete date else year only}
}
\description{
This function will download the codes of selected geographical levels via API.
}
