% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set.R
\name{get_set}
\alias{get_set}
\title{Set geo codes changes with API}
\usage{
get_set(
  type = c("fylke", "kommune", "bydel", "grunnkrets"),
  year = NULL,
  from = NULL
)
}
\arguments{
\item{type}{Type of regional granularity ie. fylke, kommune etc.}

\item{year}{Year of the selected code list}

\item{from}{Year from}
}
\description{
This is a wrapper function to get geo code list and code changes. But code changes
will only valid for every two years. To find codes changes for multiple years, use
\code{get_change} function.
}
\details{
The following dataset will be generated:
\itemize{
\item data   : The whole merged dataset
\item change : The code changes
\item type   : The granularity level
}
}
