% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_geo.R
\name{merge_geo}
\alias{merge_geo}
\title{Merge all geo codes}
\usage{
merge_geo(files)
}
\arguments{
\item{files}{A list of objects to be merged. The order must be from lowest to highest year.}
}
\description{
Create a table with all geo change and year. The files for the argument must be
the output object from \code{add_change()} function.
}
\examples{

\dontrun{
DT <- merge_geo(list(kommune2018, kommune2019, kommune2020))
}

}
