% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_change.R
\name{merge_multi}
\alias{merge_multi}
\title{Multiple changes}
\usage{
merge_multi(newCode = NULL, preCode = NULL, raw = TRUE)
}
\arguments{
\item{newCode}{Object produced from \code{geo_set} for newer codes}

\item{preCode}{Object produced from \code{geo_set} for previous codes}

\item{raw}{If FALSE then specify a data.table object}
}
\description{
Track codes that have changed multiple times. Join changes of geo codes from previous change ie.
code that have changed in 2018 # and again have new changes in 2020. Then get the previous codes
in 2018 from previous code columns.
For example the code 30240317 (in 2020) was 2190317 (2019) but previously was 2190314 (2018)
}
