% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgeo.R
\docType{package}
\name{norgeo-package}
\alias{norgeo}
\alias{norgeo-package}
\title{norgeo : A package for tracking geo codes changes in Norway}
\description{
This package tracks all geo codes changes in Norway that are available from SSB.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/helseprofil/norgeo}
  \item Report bugs at \url{https://github.com/helseprofil/norgeo/issues}
}

}
\author{
\strong{Maintainer}: Yusman Kamaleri \email{ybkamaleri@gmail.com}

}
\keyword{internal}
