% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\docType{data}
\name{raw}
\alias{raw}
\title{Where raw data will be kept}
\format{
An object of class \code{environment} of length 0.
}
\usage{
raw
}
\description{
Created object when running \code{norgeo::read_csv()}
}
\keyword{datasets}
