% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_setup.R
\name{read_ssb}
\alias{read_ssb}
\title{Read raw CSV files from SSB to R object}
\usage{
read_ssb(
  folder.path,
  type = NULL,
  sep = c(";", "."),
  encoding = NULL,
  global = FALSE
)
}
\arguments{
\item{folder.path}{Sti til mappen hvor filene ligger}

\item{type}{Hvilket nivå f.eks fylke, kommune osv.}

\item{sep}{The separate symbols in the CSV file}

\item{encoding}{Select the suitable encoding}

\item{global}{If TRUE the objects will be available globally}
}
\description{
Read all CSV files in the folder specified. The objects can be available
in the .GlobalEnv if argument \code{global = TRUE}. \strong{Important!} Any existing
objects of similar names will be replaced without warning.
The object name will be something like "kommune_01" if
\code{kommune} is specified in the \code{type} argument.
}
