% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_setup.R
\name{select_ssb}
\alias{select_ssb}
\title{Select files}
\usage{
select_ssb(grep.file = NULL, grep.change = NULL, folder.path = NULL)
}
\arguments{
\item{grep.file}{Common file name for the selected period eg. "jan2019"}

\item{grep.change}{Common name for \code{xlsx} file for code change eg. "change"}

\item{folder.path}{Folder path to the files}
}
\description{
Use \code{regex} to select files
}
