% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-merge.R
\name{track_merge}
\alias{track_merge}
\title{Get geo code that are merged after code change}
\usage{
track_merge(
  type = c("fylke", "kommune", "bydel", "grunnkrets"),
  from = NULL,
  to = NULL
)
}
\arguments{
\item{type}{Type of regional granularity ie. fylke, kommune etc.}

\item{from}{Specify the starting year for range period. Current year is the default.}

\item{to}{Specify the year to end the range period. Current year is used when not specified.}
}
\value{
Dataset with column 'merge' showing how many the codes have been merged to
}
\description{
Get geo code that are merged after code change
}
