% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-split.R
\name{track_split}
\alias{track_split}
\title{Get geo code that are split after code change}
\usage{
track_split(
  type = c("fylke", "kommune", "bydel", "grunnkrets"),
  from = NULL,
  to = NULL
)
}
\arguments{
\item{type}{Type of regional granularity ie. fylke, kommune etc.}

\item{from}{Specify the starting year for range period. Current year is the default.}

\item{to}{Specify the year to end the range period. Current year is used when not specified.}
}
\value{
Dataset with column 'split' showing how many the codes have been split to
}
\description{
Get geo code that are split after code change
}
