% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_change.R
\name{add_change}
\alias{add_change}
\title{Legg til geokoder endringer}
\usage{
add_change(
  grep.file = NULL,
  grep.change = NULL,
  year = NULL,
  type = c("land", "fylke", "kommune", "bydel", "grunnkrets"),
  folder.path = NULL,
  filegeo = NULL,
  filechg = NULL,
  raw = FALSE
)
}
\arguments{
\item{grep.file}{Felles navn til filene. Funker med bare noen ord eg. "jan2019"}

\item{grep.change}{Navn til endringsfil eg. "change"}

\item{year}{Året geokoder gjelder}

\item{type}{Hvilket nivå f.eks fylke, kommune osv.}

\item{folder.path}{Sti til mappen hvor filene ligger}

\item{filegeo}{CSV fil for alle geokoder}

\item{filechg}{Excel fil for kode endringer}

\item{raw}{If TRUE then specify direct filename in \code{filegeo} and \code{filechg}}
}
\description{
Filen for gjeldende geokoder i \code{csv} format og filen om koder som har
endret i \code{xlsx} skal legges sammen. Disse filene hentes fra SSB
hjemmeside.
}
\details{
Følgende dataset produseres:
\itemize{
\item DT     : Hele datasettet med kode endringer
\item xl     : Endringer fra Excel filen
\item change : Sti til fil endringe hvis \code{raw = FALSE}
\item code   : Sti til kodefilen hvis \code{raw = FALSE}
}
}
