% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast-geo.R
\name{cast_geo}
\alias{cast_geo}
\title{Cast geo granularity from API}
\usage{
cast_geo(year = NULL)
}
\arguments{
\item{year}{Which year the codes are valid from. If NULL then current year
will be selected.}
}
\value{
A dataset of class \code{data.table} representing the spreading of
different geographical levels from lower to higher levels ie. from
enumeration area codes to county codes, for the selected year.
}
\description{
Add geo granularity levels to all sides
}
\examples{
\dontrun{
 DT <- cast_geo(2020)
}
}
