% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast-geo.R
\name{cast_geo}
\alias{cast_geo}
\title{Cast geo granularity from API}
\usage{
cast_geo(year = NULL)
}
\arguments{
\item{year}{Which year the codes are valid from. If NULL then current year
will be selected.}
}
\description{
Add geo granularity levels to all sides
}
