% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{cast_geo}
\alias{cast_geo}
\title{Add lower granularity when applicable}
\usage{
cast_geo(file, type, year, folder.path, keep.col = c("code", "name"))
}
\arguments{
\item{file}{File names}

\item{type}{Hvilket nivå f.eks fylke, kommune osv.}

\item{year}{Året geokoder gjelder}

\item{folder.path}{Sti til mappen hvor filene ligger}

\item{keep.col}{Exisiting column names to be kept}
}
\description{
This function will add lower granularity to the current granularity when applicable.
It means that when casting a kommune file, then columns for fylke will be added.
For grunnkrets file then both kommune and fylke code will be added. This is based on
the available table that can be downloaded from SSB. The file should be in a CSV format.
}
\examples{

\dontrun{
file = "ssb_grunnkrets_jan2020.csv"
type = "grunnkrets"
year = 2020
keep.col = c("code", "name")
folder = "C:/Users/geo/grunnkrets"

DT <- cast_geo(file = file, type = type, year = year, folder.path = folder.path)
}

}
