% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast-geo.R
\name{find_correspond}
\alias{find_correspond}
\title{Find existing correspond}
\usage{
find_correspond(type, correspond, from)
}
\arguments{
\item{type}{Higher granularity from specified correspond arg.}

\item{correspond}{Lower granularity from the specified type arg.}

\item{from}{Specify the starting year for range period. Current year is the
default.}
}
\value{
A dataset of class \code{data.table} representing the lower geographical
level codes and their corresponding higher geographical levels. For example
for codes on enumeration areas and their corresponding codes for
municipalities or town.
}
\description{
Unlike \code{\link[=get_correspond]{get_correspond()}} functions, this function will find existing
correspond if the specified year has no correspond codes.
Correspond codes can be empty if nothing has changed in
that specific year and need to get from previous year or even
year before before previous year etc..etc.. This function is needed
when running \code{\link[=cast_geo]{cast_geo()}}.
}
