% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_set.R
\name{geo_set}
\alias{geo_set}
\title{Set geo codes changes}
\usage{
geo_set(type = c("fylke", "kommune", "bydel", "grunnkrets"), year = NULL)
}
\arguments{
\item{type}{Type of regional granularity ie. fylke, kommune etc.}

\item{year}{Year of the selected code list}
}
\description{
This is a wrapper function to get geo code list and code changes.
}
\details{
The following dataset will be generated:
\itemize{
\item data   : The whole merged dataset
\item xl     : The code changes
\item type   : The granularity level
}
}
