% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_set.R
\name{geo_set}
\alias{geo_set}
\title{Set geo codes changes}
\usage{
geo_set(
  grep.file = NULL,
  grep.change = NULL,
  year = NULL,
  type = c("land", "fylke", "kommune", "bydel", "grunnkrets"),
  folder.path = NULL,
  filegeo = NULL,
  filechg = NULL,
  raw = FALSE
)
}
\arguments{
\item{grep.file}{Felles navn til filene. Funker med bare noen ord eg. "jan2019"}

\item{grep.change}{Navn til endringsfil eg. "change"}

\item{year}{Året geokoder gjelder}

\item{type}{Hvilket nivå f.eks fylke, kommune osv.}

\item{folder.path}{Sti til mappen hvor filene ligger}

\item{filegeo}{CSV fil for alle geokoder}

\item{filechg}{Excel fil for kode endringer}

\item{raw}{If TRUE then specify direct filename in \code{filegeo} and \code{filechg}}
}
\description{
The downloaded file of geo codes for respective granularity from SSB in \code{csv}
format and codes that have changed copied from \emph{Endringer} tab and pasted into
an Excel file. The \code{.csv} and \code{.xlsx} files will than be merged by running this function.
}
\details{
Følgende dataset produseres:
\itemize{
\item DT     : The whole merged dataset
\item xl     : The xlsx file
\item change : Path where the change file is if \code{raw = FALSE}
\item code   : Path where the \code{csv} file is if \code{raw = FALSE}
\item type   : The granularity level
}
}
