% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_set.R
\name{geo_set}
\alias{geo_set}
\title{Set geo codes changes}
\usage{
geo_set(
  grep.file = NULL,
  grep.change = NULL,
  year = NULL,
  type = c("fylke", "kommune", "bydel", "grunnkrets"),
  folder.path = NULL,
  filegeo = NULL,
  filechg = NULL
)
}
\arguments{
\item{grep.file}{Common file name for the selected period eg. "jan2019"}

\item{grep.change}{Common name for \code{xlsx} file for code change eg. "change"}

\item{year}{Year of the selected code list}

\item{type}{Type of regional granularity ie. fylke, kommune etc.}

\item{folder.path}{Folder path to the files}

\item{filegeo}{Specify \code{csv} file for geocode if grep.file is empty}

\item{filechg}{Specify \code{xlsx} file name and location if \code{grep.change} is empty}
}
\description{
The downloaded file of geo codes for specific year from SSB is saved in \code{csv}
format. The codes that have changed are copied from \emph{Endringer} tab and pasted into
an Excel file. The \code{.csv} and \code{.xlsx} files will than be merged by running this function.
}
\details{
The following dataset will be generated:
\itemize{
\item data   : The whole merged dataset
\item xl     : The xlsx file
\item change : Path where the change file is if \code{raw = FALSE}
\item code   : Path where the \code{csv} file is if \code{raw = FALSE}
\item type   : The granularity level
}
}
