% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-change.R
\name{get_change}
\alias{get_change}
\title{Get geo code changes}
\usage{
get_change(
  type = c("fylke", "kommune", "bydel", "grunnkrets"),
  from = NULL,
  to = NULL,
  code = TRUE,
  quiet = FALSE,
  date = FALSE
)
}
\arguments{
\item{type}{Type of regional granularity ie. fylke, kommune etc.}

\item{from}{Specify the starting year for range period. Current year is the default.}

\item{to}{Specify the year to end the range period. Current year is used when not specified.}

\item{code}{TRUE will only track code changes. Else change name only will also be considered as change.}

\item{quiet}{TRUE will suppress messages when no changes happened for a specific time range}

\item{date}{If TRUE then give complete date else year only}
}
\description{
This function will download all registered geo code changes via API from SSB.
}
