% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-change.R
\name{get_change}
\alias{get_change}
\title{Get geo code changes}
\usage{
get_change(
  type = c("fylke", "kommune", "bydel", "grunnkrets"),
  year = NULL,
  from = NULL
)
}
\arguments{
\item{type}{Type of regional granularity ie. fylke, kommune etc.}

\item{year}{Year of the selected code list}

\item{from}{Year from}
}
\description{
This function will download all registered geo code changes via API from SSB.
}
