% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-change.R
\name{get_change}
\alias{get_change}
\title{Get geo code changes}
\usage{
get_change(
  type = c("fylke", "kommune", "bydel", "grunnkrets"),
  from = NULL,
  to = NULL,
  quiet = FALSE,
  date = FALSE
)
}
\arguments{
\item{type}{Higer granularity from specified correspond arg.}

\item{from}{Specify the starting year for range period. Current year is the default.}

\item{to}{Specify the year to end the range period. Indefinite is used when not specified.}

\item{quiet}{TRUE will suppress messages when no changes happened for a specific time range}

\item{date}{If TRUE then give complete date else year only}
}
\description{
This function will download all registered geo code changes via API from SSB.
}
