% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_change.R
\name{get_change}
\alias{get_change}
\title{Get change from complete geo list}
\usage{
get_change(
  files,
  years,
  type,
  folder.path,
  save = c("no", "xls", "csv"),
  des.path
)
}
\arguments{
\item{files}{A list of files to find geo codes changes}

\item{years}{A list of years for these files}

\item{type}{Type of regional granularity ie. fylke, kommune etc.}

\item{folder.path}{Folder path to the files}

\item{save}{Which format to use to save the output}

\item{des.path}{Destination folder where the file to be saved}
}
\description{
Create table for code changes when table tracking the changes isn't already available from SSB.
This function will be using a reference colum other than geo code ie. names. Therefore this
function should be used with caution. The output can be save as \code{xlsx} or \code{csv}.
}
\examples{

\dontrun{
files = c("ssb_bydel_jan2004.csv", "ssb_bydel_jan2018.csv", "ssb_bydel_jan2020.csv")
years = c(2004, 2018, 2020)
folder = "C:/geo/bydel"
des = "C:/geo/bydel/output"

get_change(files=files, years=years, type="bydel",folder.path=folder, save="xls", des.path=des)
}

}
