% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_setup.R
\name{read_csv}
\alias{read_csv}
\title{Read raw CSV files downloaded from SSB to R object}
\usage{
read_csv(
  folder.path,
  type = NULL,
  sep = c(NA, ";", ","),
  encoding = NULL,
  raw = FALSE
)
}
\arguments{
\item{folder.path}{Folder path to the files}

\item{type}{Type of regional granularity ie. fylke, kommune etc.}

\item{sep}{The separate symbols in the CSV files ie. \verb{;} or \verb{,}. If you leave it blank then
appropriate symbols will be selected automatically. But if you should specify \code{encoding} type
then, you have to choose the right \code{sep} for the file.}

\item{encoding}{Select the suitable encoding}

\item{raw}{If TRUE the objects will be available in \code{raw} environment}
}
\description{
Read all CSV files in the folder specified if you prefer not to use \code{apply} family.
The objects can be available in the \code{raw} if argument \code{raw = TRUE}.
\strong{Important!} Any existing objects of similar names will be replaced without warning.
The object name will be something like "raw$kommune200825_01" if granularity type is
\code{kommune} specified in the \code{type} argument.
}
