% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_setup.R
\name{select_ssb}
\alias{select_ssb}
\title{Select files}
\usage{
select_ssb(grep.file = NULL, grep.change = NULL, folder.path = NULL)
}
\arguments{
\item{grep.file}{Felles navn til filene. Funker med bare noen ord eg. "jan2019"}

\item{grep.change}{Navn til endringsfil eg. "change"}

\item{folder.path}{Sti til mappen hvor filene ligger}
}
\description{
Use \code{regex} to select files
}
